@@ -1,7 +1,11 @@
 #!/usr/bin/env node
 
-const yargs = require('yargs');
+import Yargs from 'yargs';
 
+import * as Install from './install.js';
+import * as Uninstall from './uninstall.js';
+
+const yargs = Yargs();
 function setCmdName(name, cmd) {
   return {
     ...cmd,
@@ -16,6 +20,6 @@ yargs
   .alias(`v`, `version`)
   .wrap(yargs.terminalWidth())
   .strict()
-  .command(setCmdName('install', require('./install')))
-  .command(setCmdName('uninstall', require('./uninstall')))
+  .command(setCmdName('install', Install))
+  .command(setCmdName('uninstall', Uninstall))
   .parse(process.argv.slice(2));
