@@ -10,4 +10,4 @@ const typescriptTransformer = (error) => {
     : error;
 };
 
-module.exports = typescriptTransformer;
+export default typescriptTransformer;
